crawler/requirements.txt
requests
feedparser
newspaper3k
beautifulsoup4
lxml
pandas
scikit-learn
